+++ viewer/src/viewer/controller/mod.rs
@@ -2,3 +2,5 @@ pub mod gizmo;
 pub use gizmo::*;
 pub mod interactive;
 pub use interactive::*;
+pub mod view;
+pub use view::*;
